+++ src/espIdf/debugAdapter/verifyApp.ts
@@ -21,12 +21,13 @@ import { l10n, Uri } from "vscode";
 import { createFlashModel } from "../../flash/flashModelBuilder";
 import { readParameter, readSerialPort } from "../../idfConfiguration";
 import { Logger } from "../../logger/logger";
-import { appendIdfAndToolsToPath, spawn } from "../../utils";
+import { spawn } from "../../utils";
 import { pathExists } from "fs-extra";
 import { getVirtualEnvPythonPath } from "../../pythonManager";
+import { configureEnvVariables } from "../../common/prepareEnv";
 
 export async function verifyAppBinary(workspaceFolder: Uri) {
-  const modifiedEnv = await appendIdfAndToolsToPath(workspaceFolder);
+  const modifiedEnv = await configureEnvVariables(workspaceFolder);
   const serialPort = await readSerialPort(workspaceFolder, false);
   if (!serialPort) {
     return Logger.warnNotify(
@@ -37,10 +38,9 @@ export async function verifyAppBinary(workspaceFolder: Uri) {
     );
   }
   const flashBaudRate = readParameter("idf.flashBaudRate", workspaceFolder);
-  const idfPath = readParameter("idf.espIdfPath", workspaceFolder);
-  const pythonBinPath = await getVirtualEnvPythonPath(workspaceFolder);
+  const pythonBinPath = await getVirtualEnvPythonPath();
   const esptoolPath = join(
-    idfPath,
+    modifiedEnv["IDF_PATH"],
     "components",
     "esptool_py",
     "esptool",