+++ src/views/cmakelists-editor/components/SetElement.vue
@@ -120,6 +120,7 @@ function clearError() {
 }
 
 .vscode-input {
+  min-width: 50%;
   height: 20px;
   padding: 4px 8px;
   background-color: var(--vscode-input-background);