+++ src/views/new-project/Templates.vue
@@ -118,7 +118,7 @@ function handleCreateProject() {
 </template>
 
 <style lang="scss">
-@import "../commons/espCommons.scss";
+@use "../commons/espCommons.scss" as *;
 
 /* Add a little spacing for the back button */
 .back-btn-wrapper {